Refactor destinations framework
