+++ MoonVR/Assets/DestinationTest.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 11f7f1b13af7eaa47a53778999c240e4
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
